+++ custom_addons/ak_library_management/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     'name': 'Library Management',
-    'version': '18.0.1.5.2',
+    'version': '18.0.1.5.3',
     'author': 'Jay Butani',
     'sequence': 5,
     'summary': 'Library Management Module',