+++ custom_crm/jaybutani_crm_customisation/models/stock_move.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from odoo import models, fields
+
+
+class StockMoveLine(models.Model):
+    _inherit = 'stock.move'
+
+    job_name = fields.Char(string='Job Name')